telegrambot/requirements.txt
python-telegram-bot
websocket-client
pyyaml
